Add credentials for quay.io
This adds credentials (uname/password) that should allow zuul jobs
to push to appropriate https://quay.io/airshipit projects via a
robot account.  The first image repo configured for this is:
https://quay.io/airshipit/isogen

Note that these credentials are encrypted and can only be decrypted
by zuul, on behalf of the airship/images project (i.e. the credentials
can't be successfully shared with other projects or actors).

Change-Id: Ie719e4ad8fbfb7c5e4287d4ceaa69b1dffa58e1a