Revert "debian-isogen: Remove unstable repository"
This reverts commit bca69c18937428afcf3e30a658665d810d7657b9.

Appears that unstable repository has been fixed. We should revert this
and find a solution from where to install or build  new version of
cloud-init instead of the one that comes from unstable repository.
Version of cloud-init that comes from stable repository, is old,
and doesn't support openstack network data properly

This is temporary fix, new issue will be created and address in
airshipctl repository

Change-Id: I02e95f8a4a84ba0c8238634abd3d9460caa62819