Fixing AKS creation failure
When the capz.kubeconfig file was missing, the creation of AKS cluster
by the bootstrap container would fail when retrieving its kubeconfig
file. The issue was that the container tries to delete an existing
kubeconfig file and would fail if it didn't exist.

The fix was to check if the kubeconfig file existed prior to deleting
it.

Change-Id: Ibf135a11ae1928fc69c35cacdf0e1339629027f2