Add support for appending to package install lists
By default, ansible overrides an entire variable. Defining a list of
package install overrides results in the replacement of the entire
original list.

This patchset adds support for appending to the list of repos, ubuntu
packages installed by multistrap, and packages installed in-target in
the post-install phase. It also exposes the k8s version as a more easily
configured override.

Change-Id: I9c6f5d570b1826bdb198d6482648544b5e5222fc