Update hostname files before multistrap in image-builder
This change adds to the multistrap playbook to ensure that the
hostname files are correct when multistrap installs packages. This
can be an issue when installing packages such as ssmtp. A later step
will override the hosts and hostname files so these changes will be
temporary in the overall build process.

Change-Id: I61124a4e991e4aaedadc9c6ef211b574178fb411