Split latest and commit tag publishing
After the cleanup optimization [0], the image publishing job is
now getting through publishing the latest tag, but the next
time through when building/publishing the commit tag, the
job is still running out of space. This splits this job into
two jobs, one for the latest tag and one for the commit tag,
to allow each to fit within the VM size.

Long term it would be better to just build the images once and push
both tags, but that will require refactoring all the Makefiles to
allow multiple tags to be passed.

[0]: https://review.opendev.org/c/airship/images/+/784607

Signed-off-by: Sean Eagan <seaneagan1@gmail.com>
Change-Id: I73e20e24de8a34130f869ec0ea4da4ddb48156a8