vino-builder wait for dynamic data
This patchset adds a wait loop in the entrypoint.sh for
vino-builder to wait for dynamic data to be available before
continuing with its processing. Since dynamic data will be
specific to a node (e.g., mac addresses), VINO will pass
these values via Kubernetes node annotations.

This patchset also comments out the playbook step to set
libivirt domains to running. In practice that will be done
in a later stage by VINO.

Change-Id: Ic6ec82b4aecf527ce329b4cb620418821efc0246